Implemented create event POST route - Moved EventRepository to new repository directory - Added imports and updated dependencies
