+++ src/main/java/com/TWCC/controller/EventController.java
@@ -6,11 +6,13 @@ import java.util.Optional;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
+import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RestController;
+import org.springframework.web.bind.annotation.RequestBody;
 
 import com.TWCC.data.Event;
-import com.TWCC.data.EventRepository;
 import com.TWCC.exception.InvalidRequestException;
+import com.TWCC.repository.EventRepository;
 
 @RestController
 public class EventController {
@@ -50,5 +52,12 @@ public class EventController {
     public List<Event> getEventsBeforeDate(@PathVariable String date) {
         // TODO: return events before a certain date
         return null;
+    }
+    
+    @PostMapping("/events")
+    public Event createEvent(@RequestBody Event newEvent) {
+        System.out.println("print new event");
+        System.out.println("new event: " + newEvent.toString());
+        return eventRepository.save(newEvent);
     }
 }
