+++ src/main/java/com/TWCC/repository/EventRepository.java
@@ -1,9 +1,11 @@
-package com.TWCC.data;
+package com.TWCC.repository;
 
 import java.util.List;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
+
+import com.TWCC.data.Event;
 
 @Repository
 public interface EventRepository extends JpaRepository<Event, Integer> {
