+++ src/test/java/com/TWCC/nyctodoservice/NycTodoServiceApplicationTests.java
@@ -1,4 +1,4 @@
-package com.TWCC.nyctodoservice;
+package com.TWCC;
 
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
